backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
google-generativeai
requests
faker
python-multipart
